integrations/data-quality/great-expectations/requirements.txt
pandas>=2.2
pyarrow>=16.0
great_expectations>=0.18
click>=8.1
python-dotenv>=1.0
requests>=2.31
